Move case when variadic scalar function from class to method
